Enum sciter::dom::event::EVENT_GROUPS

pub enum EVENT_GROUPS {
    HANDLE_INITIALIZATION,
    HANDLE_MOUSE,
    HANDLE_KEY,
    HANDLE_FOCUS,
    HANDLE_SCROLL,
    HANDLE_TIMER,
    HANDLE_SIZE,
    HANDLE_DRAW,
    HANDLE_DATA_ARRIVED,
    HANDLE_BEHAVIOR_EVENT,
    HANDLE_METHOD_CALL,
    HANDLE_SCRIPTING_METHOD_CALL,
    HANDLE_TISCRIPT_METHOD_CALL,
    HANDLE_EXCHANGE,
    HANDLE_GESTURE,
    HANDLE_ALL,
    SUBSCRIPTIONS_REQUEST,
}

Event groups for subscription.

Variants

HANDLE_INITIALIZATION

attached/detached

HANDLE_MOUSE

mouse events

HANDLE_KEY

key events

HANDLE_FOCUS

focus events, if this flag is set it also means that element it attached to is focusable

HANDLE_SCROLL

scroll events

HANDLE_TIMER

timer event

HANDLE_SIZE

size changed event

HANDLE_DRAW

drawing request (event)

HANDLE_DATA_ARRIVED

requested data () has been delivered

HANDLE_BEHAVIOR_EVENT

logical, synthetic events: BUTTON_CLICK, HYPERLINK_CLICK, etc., a.k.a. notifications from intrinsic behaviors

HANDLE_METHOD_CALL

behavior specific methods

HANDLE_SCRIPTING_METHOD_CALL

behavior specific methods

HANDLE_TISCRIPT_METHOD_CALL

behavior specific methods using direct tiscript::value's

HANDLE_EXCHANGE

system drag-n-drop

HANDLE_GESTURE

touch input events

HANDLE_ALL

all of them

SUBSCRIPTIONS_REQUEST

special value for getting subscription flags