Struct sciter::dom::Element

pub struct Element {
    // some fields omitted
}

DOM element wrapper. See the module-level documentation also.

Methods

impl Element

fn from(he: HELEMENT) -> Element

Construct Element object from HELEMENT handle.

fn create(tag: &str) -> Result<Element>

Create new element, it is disconnected initially from the DOM.

fn create_at(tag: &str, parent: &mut Element) -> Result<Element>

Create new element as child of parent.

fn with_text(tag: &str, text: &str) -> Result<Element>

Create new element with specified text, it is disconnected initially from the DOM.

fn with_type(tag: &str, el_type: &str) -> Result<Element>

Create new element with specified type, which is useful for controls and widgets (initially disconnected).

fn from_window(hwnd: HWINDOW) -> Result<Element>

Get root DOM element of the Sciter document.

fn from_focus(hwnd: HWINDOW) -> Result<Element>

Get focus DOM element of the Sciter document.

fn from_highlighted(hwnd: HWINDOW) -> Result<Element>

Get highlighted element.

fn from_point(hwnd: HWINDOW, pt: POINT) -> Result<Element>

Find DOM element of the Sciter document by coordinates.

fn from_uid(hwnd: HWINDOW, uid: u32) -> Result<Element>

Get element handle by its UID.

fn as_ptr(&self) -> HELEMENT

Access element pointer.

fn get_uid(&self) -> u32

Get element UID - identifier suitable for storage.

fn get_tag(&self) -> String

Return element tag as string (e.g. 'div', 'body').

fn get_text(&self) -> String

Get inner text of the element as string.

fn set_text(&mut self, text: &str) -> Result<()>

Set inner text of the element.

fn get_html(&self, with_outer_html: bool) -> Vec<u8>

Get html representation of the element as utf-8 bytes.

fn set_html(&mut self, html: &[u8], how: Option<SET_ELEMENT_HTML>) -> Result<()>

Set inner or outer html of the element.

fn get_value(&self) -> Value

Get value of the element.

fn set_value(&mut self, val: Value) -> Result<()>

Set value of the element.

fn get_hwnd(&self, for_root: bool) -> HWINDOW

Get HWINDOW of containing window.

fn eval_script(&self, script: &str) -> Result<Value>

Evaluate script in element context.

fn call_function(&self, name: &str, args: &[Value]) -> Result<Value>

Call scripting function defined in the namespace of the element (a.k.a. global function).

You can use the make_args!(a,b,c) macro which help you construct script arguments from Rust types.

fn call_method(&self, name: &str, args: &[Value]) -> Result<Value>

Call scripting method defined for the element.

You can use the make_args!(a,b,c) macro which help you construct script arguments from Rust types.

fn attribute_count(&self) -> usize

Get number of the attributes.

fn attribute_name(&self, index: usize) -> String

Get attribute name by its index.

fn attribute(&self, index: usize) -> String

Get attribute value by its index.

fn get_attribute(&self, name: &str) -> Option<String>

Get attribute value by its name.

fn set_attribute(&mut self, name: &str, value: &str) -> Result<()>

Add or replace attribute.

fn remove_attribute(&mut self, name: &str) -> Result<()>

Remove attribute.

fn toggle_attribute(&mut self, name: &str, isset: bool, value: Option<&str>) -> Result<()>

Toggle attribute.

fn clear_attributes(&mut self) -> Result<()>

Remove all attributes from the element.

fn get_style_attribute(&self, name: &str) -> String

Get style attribute of the element by its name.

fn set_style_attribute(&mut self, name: &str, value: &str) -> Result<()>

Set style attribute.

fn index(&self) -> usize

Get index of this element in its parent collection.

fn root(&self) -> Element

Get root of the element.

fn parent(&self) -> Option<Element>

Get parent element.

fn first_sibling(&self) -> Option<Element>

Get first sibling element.

fn last_sibling(&self) -> Option<Element>

Get last sibling element.

fn next_sibling(&self) -> Option<Element>

Get next sibling element.

fn prev_sibling(&self) -> Option<Element>

Get previous sibling element.

fn first_child(&self) -> Option<Element>

Get first child element.

fn last_child(&self) -> Option<Element>

Get last child element.

fn get(&self, index: usize) -> Option<Element>

Get element child at specified index.

fn child(&self, index: usize) -> Option<Element>

Get element child at specified index.

fn children_count(&self) -> usize

Get number of child elements.

fn len(&self) -> usize

Get number of child elements.

fn clear(&mut self) -> Result<()>

Clear content of the element.

fn clone_element(&self) -> Element

Create new element as copy of existing element.

The new element is a full (deep) copy of the element and is initially disconnected from the DOM. Note that Element.clone() does not clone DOM element, just increments its reference count.

fn insert(&mut self, index: usize, child: &Element) -> Result<()>

Insert element at index position of this element.

Note that we cannot follow Rust semantic here because the newly created Element is unusable before it will be inserted at DOM.

fn append(&mut self, child: &Element) -> Result<()>

Append element as last child of this element.

fn push(&mut self, element: Element)

Append element as last child of this element.

fn pop(&mut self) -> Option<Element>

Remove the last child from this element and returns it, or None if this element is empty.

fn detach(&mut self) -> Result<()>

Take element out of its container (and DOM tree).

fn destroy(&mut self) -> Result<()>

Take element out of its container (and DOM tree) and force destruction of all behaviors.

fn swap(&mut self, other: &mut Element) -> Result<()>

Swap element positions.

fn test(&self, selector: &str) -> bool

Test this element against CSS selector(s).

fn find_nearest_parent(&self, selector: &str) -> Result<Option<Element>>

Will find first parent element starting from this satisfying given css selector(s).

fn find_first(&self, selector: &str) -> Result<Option<Element>>

Will find first element starting from this satisfying given css selector(s).

fn find_all(&self, selector: &str) -> Result<Option<Vec<Element>>>

Will find all elements starting from this satisfying given css selector(s).

fn update(&self, render_now: bool) -> Result<()>

Apply changes and refresh element area in its window.

fn start_timer(&self, period_ms: u32, timer_id: u64) -> Result<()>

Start Timer for the element. Element will receive on_timer event.

Note that timer events are not bubbling, so you need attach handler to the target element directly.

fn stop_timer(&self, timer_id: u64) -> Result<()>

Stop Timer for the element.

fn attach_handler<T: EventHandler>(&mut self, handler: T) -> Result<u64>

Attach the native event handler to this element.

fn detach_handler<T: EventHandler>(&mut self, token: u64) -> Result<()>

Detach your handler from the element. Handlers identified by token from attach_handler() result.