Module sciter::host

Sciter host application helpers.

Structs

Host
Sciter host runtime support.

SCN_ATTACH_BEHAVIOR
This notification is sent on parsing the document and while processing elements having non empty style.behavior attribute value.

SCN_DATA_LOADED
This notification indicates that external data (for example image) download process completed.

SCN_LOAD_DATA
Notifies that Sciter is about to download a referred resource.

Enums

GFX_LAYER
Explicitly set sciter graphics layer (Windows only).

LOAD_RESULT
HostHandler::on_data_load() result code.

OUTPUT_SEVERITY
Debug output severity.

OUTPUT_SUBSYTEMS
Debug output categories.

SCITER_RT_OPTIONS
Various sciter engine options.

SCRIPT_RUNTIME_FEATURES
Script runtime options.

Traits

HostHandler
Sciter notification handler for Window.sciter_handler().

Type Definitions

Result
A specialized Result type for sciter host operations.