Struct sciter::host::Host

pub struct Host {
    // some fields omitted
}

Sciter host runtime support.

Methods

impl Host

fn setup_callback<T: HostHandler>(&self, hwnd: HWINDOW, handler: T)

Set callback for sciter engine events.

fn enable_debug(&self, enable: bool)

Setup debug output function for specific window or globally.

fn get_hwnd(&self) -> HWINDOW

Get native window handle.

fn get_root(&self) -> Option<Element>

Get window root DOM element.

fn load_file(&self, uri: &str)

Load HTML document from file.

fn load_html(&self, html: &[u8], uri: Option<&str>)

Load HTML document from memory.

fn data_ready(&self, uri: &str, data: &[u8], request_id: Option<HREQUEST>)

This function is used as response to SC_LOAD_DATA request.

fn eval_script(&self, script: &str) -> Result<Value, Value>

Evaluate script in context of current document.

fn call_function(&self, name: &str, args: &[Value]) -> Result<Value, Value>

Call scripting function defined in the global namespace.

This function returns Result<Value,Value> with script function result value or with sciter script error. You can use the make_args!(a,b,c) macro which help you construct script arguments from Rust types.

fn set_option(&self, option: SCITER_RT_OPTIONS, value: usize) -> Result<()>

Set various sciter engine options, see the SCITER_RT_OPTIONS.

fn set_home_url(&self, url: &str) -> Result<()>

Set home url for sciter resources.

If you will set it like set_home_url("http://sciter.com/modules/") then

<script src="sciter:lib/root-extender.tis"> will load root-extender.tis from

http://sciter.com/modules/lib/root-extender.tis.

fn set_media_type(&self, media_type: &str) -> Result<()>

Set media type of this sciter instance.

fn set_media_vars(&self, media: Value) -> Result<()>

Set media variables (dictionary) for this sciter instance.

By default sciter window has "screen:true" and "desktop:true"/"handheld:true" media variables.

Media variables can be changed in runtime. This will cause styles of the document to be reset.

fn set_master_css(&self, css: &str, append: bool) -> Result<()>

Set or append the master style sheet styles (globally).

fn set_window_css(&self, css: &str, base_url: &str, media_type: &str) -> Result<()>

Set (reset) style sheet of current document.