Struct sciter::host::SCN_ATTACH_BEHAVIOR

pub struct SCN_ATTACH_BEHAVIOR {
    pub code: UINT,
    pub hwnd: HWINDOW,
    pub element: HELEMENT,
    pub name: LPCSTR,
    pub elementProc: *mut ElementEventProc,
    pub elementTag: LPVOID,
}

This notification is sent on parsing the document and while processing elements having non empty style.behavior attribute value.

Fields

code

SC_ATTACH_BEHAVIOR here.

hwnd

HWINDOW of the window this callback was attached to.

element

[in] target DOM element handle

name

[in] zero terminated string, string appears as value of CSS behavior: attribute.

elementProc

[out] pointer to ElementEventProc function.

elementTag

[out] tag value, passed as is into pointer ElementEventProc function.