Struct sciter::host::SCN_DATA_LOADED

pub struct SCN_DATA_LOADED {
    pub code: UINT,
    pub hwnd: HWINDOW,
    pub uri: LPCWSTR,
    pub data: LPCBYTE,
    pub dataSize: UINT,
    pub dataType: UINT,
    pub status: UINT,
}

This notification indicates that external data (for example image) download process completed.

Fields

code

SC_DATA_LOADED here.

hwnd

HWINDOW of the window this callback was attached to.

uri

[in] zero terminated string, fully qualified uri, for example "http://server/folder/file.ext".

data

[in] pointer to loaded data.

dataSize

[in] loaded data size (in bytes).

dataType

[in] resource type category

status

Download status code:

status = 0 and dataSize == 0 - unknown error.

status = 100..505 - http response status, note: 200 - OK!

status > 12000 - wininet error code, see ERROR_INTERNET_*** in wininet.h