Struct sciter::host::SCN_LOAD_DATA

pub struct SCN_LOAD_DATA {
    pub code: UINT,
    pub hwnd: HWINDOW,
    pub uri: LPCWSTR,
    pub outData: LPCBYTE,
    pub outDataSize: UINT,
    pub dataType: SCITER_RESOURCE_TYPE,
    pub request_id: HREQUEST,
    pub principal: HELEMENT,
    pub initiator: HELEMENT,
}

Notifies that Sciter is about to download a referred resource.

Fields

code

SC_LOAD_DATA here.

hwnd

HWINDOW of the window this callback was attached to.

uri

[in] Zero terminated string, fully qualified uri, for example "http://server/folder/file.ext".

outData

[in,out] pointer to loaded data to return. If data exists in the cache then this field contain pointer to it.

outDataSize

[in,out] loaded data size to return.

dataType

[in] resource type category

request_id

[in] request handle that can be used with sciter request API.

principal

[in] destination element for request.

initiator

[in] source element for request.