Module sciter::types

Export platform-dependent types used by Sciter.

Structs

METHOD_PARAMS

MSG

POINT

RECT

REQUEST_PARAM

SCITER_CALLBACK_NOTIFICATION

SCN_ATTACH_BEHAVIOR
This notification is sent on parsing the document and while processing elements having non empty style.behavior attribute value.

SCN_DATA_LOADED
This notification indicates that external data (for example image) download process completed.

SCN_LOAD_DATA
Notifies that Sciter is about to download a referred resource.

SIZE

_HSARCHIVE

_HWINDOW

Enums

GFX_LAYER
Explicitly set sciter graphics layer (Windows only).

ID2D1Factory

ID2D1RenderTarget

IDWriteFactory

IDXGISurface

IDXGISwapChain

LOAD_RESULT
HostHandler::on_data_load() result code.

OUTPUT_SEVERITY
Debug output severity.

OUTPUT_SUBSYTEMS
Debug output categories.

SCITER_CREATE_WINDOW_FLAGS
Window flags

SCITER_NOTIFICATION

SCITER_RESOURCE_TYPE
Sciter resource type identifiers.

SCITER_RT_OPTIONS
Various sciter engine options.

SCRIPT_RUNTIME_FEATURES
Script runtime options.

Type Definitions

BOOL

BYTE

CHAR

DEBUG_OUTPUT_PROC

ElementEventProc

FLOAT_VALUE

HSARCHIVE

HWINDOW

INT

INT64

KeyValueCallback

LONG

LPARAM

LPCBYTE

LPCBYTE_RECEIVER

LPCRECT

LPCSTR

LPCSTR_RECEIVER

LPCVOID

LPCWSTR

LPCWSTR_RECEIVER

LPMSG

LPPOINT

LPRECT

LPSCITER_CALLBACK_NOTIFICATION

LPSIZE

LPSTR

LPUINT

LPVOID

LPWSTR

LRESULT

PBOOL

SciterHostCallback

SciterWindowDelegate

UINT

UINT64

UINT_PTR

VOID

WCHAR

WPARAM