Enum sciter::types::LOAD_RESULT

pub enum LOAD_RESULT {
    LOAD_DEFAULT,
    LOAD_DISCARD,
    LOAD_DELAYED,
    LOAD_MYSELF,
}

HostHandler::on_data_load() result code.

This notification gives application a chance to override built-in loader and implement loading of resources in its own way (for example images can be loaded from database or other resource).

Variants

LOAD_DEFAULT

Do default loading if data not set.

LOAD_DISCARD

Discard request completely (data will not be loaded at document).

LOAD_DELAYED

Data will be delivered later by the host application.

LOAD_MYSELF

You return this result to indicate that your (the host) application took or will take care about HREQUEST in your code completely.