Struct sciter::types::SCN_ATTACH_BEHAVIOR

#[repr(C)]

pub struct SCN_ATTACH_BEHAVIOR {
    pub code: UINT,
    pub hwnd: HWINDOW,
    pub element: HELEMENT,
    pub name: LPCSTR,
    pub elementProc: *mut ElementEventProc,
    pub elementTag: LPVOID,
}

This notification is sent on parsing the document and while processing elements having non empty style.behavior attribute value.

Fields
code: UINT

SC_ATTACH_BEHAVIOR here.

hwnd: HWINDOW

HWINDOW of the window this callback was attached to.

element: HELEMENT

[in] target DOM element handle

name: LPCSTR

[in] zero terminated string, string appears as value of CSS behavior: attribute.

elementProc: *mut ElementEventProc

[out] pointer to ElementEventProc function.

elementTag: LPVOID

[out] tag value, passed as is into pointer ElementEventProc function.