Struct sciter::types::SCN_LOAD_DATA

#[repr(C)]

pub struct SCN_LOAD_DATA {
    pub code: UINT,
    pub hwnd: HWINDOW,
    pub uri: LPCWSTR,
    pub outData: LPCBYTE,
    pub outDataSize: UINT,
    pub dataType: SCITER_RESOURCE_TYPE,
    pub request_id: *mut _HREQUEST,
    pub principal: HELEMENT,
    pub initiator: HELEMENT,
}

Notifies that Sciter is about to download a referred resource.

Fields
code: UINT

SC_LOAD_DATA here.

hwnd: HWINDOW

HWINDOW of the window this callback was attached to.

uri: LPCWSTR

[in] Zero terminated string, fully qualified uri, for example "http://server/folder/file.ext".

outData: LPCBYTE

[in,out] pointer to loaded data to return. If data exists in the cache then this field contain pointer to it.

outDataSize: UINT

[in,out] loaded data size to return.

dataType: SCITER_RESOURCE_TYPE

[in] resource type category

request_id: *mut _HREQUEST

[in] request handle that can be used with sciter request API.

principal: HELEMENT

[in] destination element for request.

initiator: HELEMENT

[in] source element for request.