Struct sciter::value::Value

pub struct Value {
    // some fields omitted
}

sciter::value wrapper. See the module-level documentation.

Methods

impl Value

fn new() -> Value

Return a new sciter value object (undefined).

fn null() -> Value

Make explicit json null value.

fn symbol(val: &str) -> Value

Make sciter symbol value.

fn error(val: &str) -> Value

Make sciter error value.

fn parse(val: &str) -> Result<Value, u32>

Parse json string into value.

fn parse_as(val: &str, how: VALUE_STRING_CVT_TYPE) -> Result<Value, u32>

Parse json string into value.

fn from_str(val: &str) -> Result<Self, VALUE_RESULT>

Parse json string into value. Note that Value::from_str() parses a json string to value object and returns a Result<Value> unlike Value::from(), which returns just string object only.

fn get_type(&self) -> VALUE_TYPE

Get inner value type.

fn full_type(&self) -> (VALUE_TYPE, UINT)

Get inner value type and its subtype (e.g. units).

fn isolate(&mut self)

Convert T_OBJECT value type to JSON T_MAP or T_ARRAY.

fn clear(&mut self) -> &mut Value

Clear the VALUE and deallocates all assosiated structures that are not used anywhere else.

fn len(&self) -> usize

Return the number of items in the T_ARRAY, T_MAP, T_FUNCTION and T_OBJECT sciter::value.

fn push(&mut self, src: Value)

Append value to the end of T_ARRAY sciter::value.

fn set(&mut self, index: usize, src: Value)

Insert or set value at given index of T_ARRAY, T_MAP, T_FUNCTION and T_OBJECT sciter::value.

fn get(&self, index: usize) -> Value

Retreive value of sub-element at index

T_ARRAY - nth element of the array;

T_MAP - value of nth key/value pair in the map;

T_FUNCTION - value of nth argument of the function.

fn set_item(&mut self, key: Value, value: Value)

Insert or set value of sub-element by key.

if it is a map - sets named value in the map;

if it is a function - sets named argument of the function;

if it is a object - sets value of property of the object;

otherwise it converts this to map and adds key/v to it.

fn get_item(&self, key: Value) -> Value

Retrieve value of sub-element by key.

fn to_int(&self) -> Option<i32>

Value to integer.

fn to_bool(&self) -> Option<bool>

Value to bool.

fn to_float(&self) -> Option<f64>

Value to float.

fn as_string(&self) -> Option<String>

Value as string for T_STRING type.

fn into_string(&mut self) -> String

Value to json string (converted in-place). Subject to change.

fn as_bytes(&self) -> Option<&[u8]>

Value as byte slice for T_BYTES type.

fn to_bytes(&self) -> Option<Vec<u8>>

Value to byte vector for T_BYTES type.

fn call(&self, this: Option<Value>, args: &[Value], name: Option<&str>) -> Result<Value, VALUE_RESULT>

Function invocation for T_OBJECT/UT_OBJECT_FUNCTION.

Calls the tiscript function or method holded at Value with context of this object that will be known as this inside that function (it is optional for global functions). name here is an url or name of the script - used for error reporting in the script. You can use the make_args!(a,b,c) macro which help you construct script arguments from Rust types.

fn is_undefined(&self) -> bool

fn is_null(&self) -> bool

fn is_bool(&self) -> bool

fn is_int(&self) -> bool

fn is_float(&self) -> bool

fn is_bytes(&self) -> bool

fn is_string(&self) -> bool

fn is_symbol(&self) -> bool

fn is_error_string(&self) -> bool

fn is_date(&self) -> bool

fn is_currency(&self) -> bool

fn is_map(&self) -> bool

fn is_array(&self) -> bool

fn is_function(&self) -> bool

fn is_native_function(&self) -> bool

fn is_object(&self) -> bool

fn is_dom_element(&self) -> bool