Module sciter::window

High level window wrapper.

To create instance of Sciter you will need either to create new Sciter window or to attach (mix-in) Sciter engine to existing window.

Handle of the Sciter engine is defined as HWINDOW type which is:

HWND handle on Microsoft Windows.

NSView* – pointer to NSView instance that is a contentView of Sciter window on OS X.

GtkWidget* – pointer to GtkWidget instance that is a root widget of Sciter window on Linux/GTK.

Creation of new window

extern crate sciter;

fn main() {
    let mut frame = sciter::Window::new();
    frame.load_file("minimal.htm");
    frame.run_app(true);
}

Also you can register the host and DOM event handlers.

Structs

Window
Sciter window.

Enums

SCITER_CREATE_WINDOW_FLAGS
Window flags

Type Definitions

Flags
SCITER_CREATE_WINDOW_FLAGS alias.