Enum sciter::window::SCITER_CREATE_WINDOW_FLAGS

pub enum SCITER_CREATE_WINDOW_FLAGS {
    SW_CHILD,
    SW_TITLEBAR,
    SW_RESIZEABLE,
    SW_TOOL,
    SW_CONTROLS,
    SW_GLASSY,
    SW_ALPHA,
    SW_MAIN,
    SW_POPUP,
    SW_ENABLE_DEBUG,
    SW_OWNS_VM,
}

Window flags

Variants

SW_CHILD

child window only, if this flag is set all other flags ignored

SW_TITLEBAR

toplevel window, has titlebar

SW_RESIZEABLE

has resizeable frame

SW_TOOL

is tool window

SW_CONTROLS

has minimize / maximize buttons

SW_GLASSY

glassy window ( DwmExtendFrameIntoClientArea on windows )

SW_ALPHA

transparent window ( e.g. WS_EX_LAYERED on Windows )

SW_MAIN

main window of the app, will terminate the app on close

SW_POPUP

the window is created as topmost window.

SW_ENABLE_DEBUG

make this window inspector ready

SW_OWNS_VM

it has its own script VM

Methods

impl SCITER_CREATE_WINDOW_FLAGS

fn main_window(resizeable: bool) -> SCITER_CREATE_WINDOW_FLAGS

Main application window.

fn child_window() -> SCITER_CREATE_WINDOW_FLAGS

Child window.

fn popup_window(title: bool, minmaxclose: bool, resizeable: bool) -> SCITER_CREATE_WINDOW_FLAGS

Popup window.

fn tool_window(resizeable: bool) -> SCITER_CREATE_WINDOW_FLAGS

Tool window.